Change service task proxying to use ips
