+++ plugin/generator_test.go
@@ -26,397 +26,6 @@ func fmtLabel(s string) string {
 	return fmt.Sprintf(s, defaultLabelPrefix)
 }
 
-func TestAddContainerWithTemplates(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ContainersData = []types.Container{
-		types.Container{
-			Names: []string{
-				"container-name",
-			},
-			NetworkSettings: &types.SummaryNetworkSettings{
-				Networks: map[string]*network.EndpointSettings{
-					"caddy-network": &network.EndpointSettings{
-						IPAddress: "172.17.0.2",
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-			Labels: map[string]string{
-				fmtLabel("%s"):       "{{index .Names 0}}.testdomain.com",
-				fmtLabel("%s.proxy"): "/ {{(index .NetworkSettings.Networks \"caddy-network\").IPAddress}}:5000/api",
-			},
-		},
-	}
-
-	const expectedCaddyfile = "container-name.testdomain.com {\n" +
-		"  proxy / 172.17.0.2:5000/api\n" +
-		"}\n"
-
-	testGeneration(t, dockerClient, false, true, expectedCaddyfile, skipCaddyfileText)
-}
-
-func TestAddContainerPicksRightNetwork(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ContainersData = []types.Container{
-		types.Container{
-			NetworkSettings: &types.SummaryNetworkSettings{
-				Networks: map[string]*network.EndpointSettings{
-					"other-network": &network.EndpointSettings{
-						IPAddress: "10.0.0.1",
-						NetworkID: "other-network-id",
-					},
-					"caddy-network": &network.EndpointSettings{
-						IPAddress: "172.17.0.2",
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-			Labels: map[string]string{
-				fmtLabel("%s.address"): "service.testdomain.com",
-			},
-		},
-	}
-
-	const expectedCaddyfile = "service.testdomain.com {\n" +
-		"  proxy / 172.17.0.2\n" +
-		"}\n"
-
-	testGeneration(t, dockerClient, false, true, expectedCaddyfile, skipCaddyfileText)
-}
-
-func TestAddContainerWithMinimumBasicLabels(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ContainersData = []types.Container{
-		types.Container{
-			NetworkSettings: &types.SummaryNetworkSettings{
-				Networks: map[string]*network.EndpointSettings{
-					"caddy-network": &network.EndpointSettings{
-						IPAddress: "172.17.0.2",
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-			Labels: map[string]string{
-				fmtLabel("%s.address"): "service.testdomain.com",
-			},
-		},
-	}
-
-	const expectedCaddyfile = "service.testdomain.com {\n" +
-		"  proxy / 172.17.0.2\n" +
-		"}\n"
-
-	testGeneration(t, dockerClient, false, true, expectedCaddyfile, skipCaddyfileText)
-}
-
-func TestAddContainerWithAllBasicLabels(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ContainersData = []types.Container{
-		types.Container{
-			NetworkSettings: &types.SummaryNetworkSettings{
-				Networks: map[string]*network.EndpointSettings{
-					"caddy-network": &network.EndpointSettings{
-						IPAddress: "172.17.0.2",
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-			Labels: map[string]string{
-				fmtLabel("%s.address"):        "service.testdomain.com",
-				fmtLabel("%s.targetport"):     "5000",
-				fmtLabel("%s.targetpath"):     "/api",
-				fmtLabel("%s.targetprotocol"): "https",
-			},
-		},
-	}
-
-	const expectedCaddyfile = "service.testdomain.com {\n" +
-		"  proxy / https://172.17.0.2:5000/api\n" +
-		"}\n"
-
-	testGeneration(t, dockerClient, false, true, expectedCaddyfile, skipCaddyfileText)
-}
-
-func TestAddContainerFromDifferentNetwork(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ContainersData = []types.Container{
-		types.Container{
-			ID: "CONTAINER-ID",
-			NetworkSettings: &types.SummaryNetworkSettings{
-				Networks: map[string]*network.EndpointSettings{
-					"other-network": &network.EndpointSettings{
-						IPAddress: "10.0.0.1",
-						NetworkID: "other-network-id",
-					},
-				},
-			},
-			Labels: map[string]string{
-				fmtLabel("%s.address"): "service.testdomain.com",
-			},
-		},
-	}
-
-	const expectedCaddyfile = ""
-
-	const expectedLogs = skipCaddyfileText +
-		"[ERROR] Container CONTAINER-ID and caddy are not in same network\n"
-
-	testGeneration(t, dockerClient, false, true, expectedCaddyfile, expectedLogs)
-}
-
-func TestAddContainerFromDifferentNetworkSkipValidation(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ContainersData = []types.Container{
-		types.Container{
-			ID: "CONTAINER-ID",
-			NetworkSettings: &types.SummaryNetworkSettings{
-				Networks: map[string]*network.EndpointSettings{
-					"other-network": &network.EndpointSettings{
-						IPAddress: "10.0.0.1",
-						NetworkID: "other-network-id",
-					},
-				},
-			},
-			Labels: map[string]string{
-				fmtLabel("%s.address"): "service.testdomain.com",
-			},
-		},
-	}
-
-	const expectedCaddyfile = "service.testdomain.com {\n" +
-		"  proxy / 10.0.0.1\n" +
-		"}\n"
-
-	const expectedLogs = skipCaddyfileText
-
-	testGeneration(t, dockerClient, false, false, expectedCaddyfile, expectedLogs)
-}
-
-func TestAddContainerWithMultipleConfigs(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ContainersData = []types.Container{
-		types.Container{
-			NetworkSettings: &types.SummaryNetworkSettings{
-				Networks: map[string]*network.EndpointSettings{
-					"caddy-network": &network.EndpointSettings{
-						IPAddress: "172.17.0.2",
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-			Labels: map[string]string{
-				fmtLabel("%s_0.address"):    "service1.testdomain.com",
-				fmtLabel("%s_0.targetport"): "5000",
-				fmtLabel("%s_0.targetpath"): "/api",
-				fmtLabel("%s_0.tls.dns"):    "route53",
-				fmtLabel("%s_1.address"):    "service2.testdomain.com",
-				fmtLabel("%s_1.targetport"): "5001",
-				fmtLabel("%s_1.tls.dns"):    "route53",
-			},
-		},
-	}
-
-	const expectedCaddyfile = "service1.testdomain.com {\n" +
-		"  proxy / 172.17.0.2:5000/api\n" +
-		"  tls {\n" +
-		"    dns route53\n" +
-		"  }\n" +
-		"}\n" +
-		"service2.testdomain.com {\n" +
-		"  proxy / 172.17.0.2:5001\n" +
-		"  tls {\n" +
-		"    dns route53\n" +
-		"  }\n" +
-		"}\n"
-
-	testGeneration(t, dockerClient, false, true, expectedCaddyfile, skipCaddyfileText)
-}
-
-func TestAddContainerWithReplicas(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ContainersData = []types.Container{
-		types.Container{
-			NetworkSettings: &types.SummaryNetworkSettings{
-				Networks: map[string]*network.EndpointSettings{
-					"caddy-network": &network.EndpointSettings{
-						IPAddress: "172.17.0.2",
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-			Labels: map[string]string{
-				fmtLabel("%s.address"): "service.testdomain.com",
-			},
-		},
-		types.Container{
-			NetworkSettings: &types.SummaryNetworkSettings{
-				Networks: map[string]*network.EndpointSettings{
-					"caddy-network": &network.EndpointSettings{
-						IPAddress: "172.17.0.3",
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-			Labels: map[string]string{
-				fmtLabel("%s.address"): "service.testdomain.com",
-			},
-		},
-	}
-
-	const expectedCaddyfile = "service.testdomain.com {\n" +
-		"  proxy / 172.17.0.2 172.17.0.3\n" +
-		"}\n"
-
-	testGeneration(t, dockerClient, false, true, expectedCaddyfile, skipCaddyfileText)
-}
-
-func TestDoNotMergeProxiesWithDifferentLabelKey(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ContainersData = []types.Container{
-		types.Container{
-			NetworkSettings: &types.SummaryNetworkSettings{
-				Networks: map[string]*network.EndpointSettings{
-					"caddy-network": &network.EndpointSettings{
-						IPAddress: "172.17.0.2",
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-			Labels: map[string]string{
-				fmtLabel("%s"):         "service.testdomain.com",
-				fmtLabel("%s.proxy_0"): "/a service-a",
-			},
-		},
-		types.Container{
-			NetworkSettings: &types.SummaryNetworkSettings{
-				Networks: map[string]*network.EndpointSettings{
-					"caddy-network": &network.EndpointSettings{
-						IPAddress: "172.17.0.3",
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-			Labels: map[string]string{
-				fmtLabel("%s"):         "service.testdomain.com",
-				fmtLabel("%s.proxy_1"): "/b service-b",
-			},
-		},
-	}
-
-	const expectedCaddyfile = "service.testdomain.com {\n" +
-		"  proxy /a service-a\n" +
-		"  proxy /b service-b\n" +
-		"}\n"
-
-	testGeneration(t, dockerClient, false, true, expectedCaddyfile, skipCaddyfileText)
-}
-
-func TestAddContainersWithSnippets(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ContainersData = []types.Container{
-		types.Container{
-			NetworkSettings: &types.SummaryNetworkSettings{
-				Networks: map[string]*network.EndpointSettings{
-					"caddy-network": &network.EndpointSettings{
-						IPAddress: "172.17.0.3",
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-			Labels: map[string]string{
-				fmtLabel("%s.address"): "service.testdomain.com",
-				fmtLabel("%s.import"):  "mysnippet-1",
-			},
-		},
-		types.Container{
-			NetworkSettings: &types.SummaryNetworkSettings{
-				Networks: map[string]*network.EndpointSettings{
-					"caddy-network": &network.EndpointSettings{
-						IPAddress: "172.17.0.2",
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-			Labels: map[string]string{
-				fmtLabel("%s_1"):     "(mysnippet-1)",
-				fmtLabel("%s_1.tls"): "off",
-				fmtLabel("%s_2"):     "(mysnippet-2)",
-				fmtLabel("%s_2.tls"): "off",
-			},
-		},
-	}
-
-	const expectedCaddyfile = "(mysnippet-1) {\n" +
-		"  tls off\n" +
-		"}\n" +
-		"(mysnippet-2) {\n" +
-		"  tls off\n" +
-		"}\n" +
-		"service.testdomain.com {\n" +
-		"  import mysnippet-1\n" +
-		"  proxy / 172.17.0.3\n" +
-		"}\n"
-
-	testGeneration(t, dockerClient, false, true, expectedCaddyfile, skipCaddyfileText)
-}
-
-func TestAddServiceWithTemplates(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ServicesData = []swarm.Service{
-		swarm.Service{
-			Spec: swarm.ServiceSpec{
-				Annotations: swarm.Annotations{
-					Name: "service",
-					Labels: map[string]string{
-						fmtLabel("%s"):                    "{{.Spec.Name}}.testdomain.com",
-						fmtLabel("%s.proxy"):              "/ {{.Spec.Name}}:5000/api",
-						fmtLabel("%s.proxy.transparent"):  "",
-						fmtLabel("%s.proxy.health_check"): "/health",
-						fmtLabel("%s.proxy.websocket"):    "",
-						fmtLabel("%s.gzip"):               "",
-						fmtLabel("%s.basicauth"):          "/ user password",
-						fmtLabel("%s.tls.dns"):            "route53",
-						fmtLabel("%s.rewrite_0"):          "/path1 /path2",
-						fmtLabel("%s.rewrite_1"):          "/path3 /path4",
-						fmtLabel("%s.limits.header"):      "100kb",
-						fmtLabel("%s.limits.body_0"):      "/path1 2mb",
-						fmtLabel("%s.limits.body_1"):      "/path2 4mb",
-					},
-				},
-			},
-			Endpoint: swarm.Endpoint{
-				VirtualIPs: []swarm.EndpointVirtualIP{
-					swarm.EndpointVirtualIP{
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-		},
-	}
-
-	const expectedCaddyfile = "service.testdomain.com {\n" +
-		"  basicauth / user password\n" +
-		"  gzip\n" +
-		"  limits {\n" +
-		"    body /path1 2mb\n" +
-		"    body /path2 4mb\n" +
-		"    header 100kb\n" +
-		"  }\n" +
-		"  proxy / service:5000/api {\n" +
-		"    health_check /health\n" +
-		"    transparent\n" +
-		"    websocket\n" +
-		"  }\n" +
-		"  rewrite /path1 /path2\n" +
-		"  rewrite /path3 /path4\n" +
-		"  tls {\n" +
-		"    dns route53\n" +
-		"  }\n" +
-		"}\n"
-
-	testGeneration(t, dockerClient, false, true, expectedCaddyfile, skipCaddyfileText)
-}
-
 func TestUseTemplatesToGenerateEmptyValues(t *testing.T) {
 	dockerClient := createBasicDockerClientMock()
 	dockerClient.ServicesData = []swarm.Service{
@@ -452,311 +61,6 @@ func TestUseTemplatesToGenerateEmptyValues(t *testing.T) {
 		"}\n"
 
 	testGeneration(t, dockerClient, false, true, expectedCaddyfile, skipCaddyfileText)
-}
-
-func TestAddServiceWithMinimumBasicLabels(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ServicesData = []swarm.Service{
-		swarm.Service{
-			Spec: swarm.ServiceSpec{
-				Annotations: swarm.Annotations{
-					Name: "service",
-					Labels: map[string]string{
-						fmtLabel("%s.address"): "service.testdomain.com",
-					},
-				},
-			},
-			Endpoint: swarm.Endpoint{
-				VirtualIPs: []swarm.EndpointVirtualIP{
-					swarm.EndpointVirtualIP{
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-		},
-	}
-
-	const expectedCaddyfile = "service.testdomain.com {\n" +
-		"  proxy / service\n" +
-		"}\n"
-
-	testGeneration(t, dockerClient, false, true, expectedCaddyfile, skipCaddyfileText)
-}
-
-func TestAddServiceWithAllBasicLabels(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ServicesData = []swarm.Service{
-		swarm.Service{
-			Spec: swarm.ServiceSpec{
-				Annotations: swarm.Annotations{
-					Name: "service",
-					Labels: map[string]string{
-						fmtLabel("%s.address"):        "service.testdomain.com",
-						fmtLabel("%s.targetport"):     "5000",
-						fmtLabel("%s.targetpath"):     "/api",
-						fmtLabel("%s.targetprotocol"): "https",
-					},
-				},
-			},
-			Endpoint: swarm.Endpoint{
-				VirtualIPs: []swarm.EndpointVirtualIP{
-					swarm.EndpointVirtualIP{
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-		},
-	}
-
-	const expectedCaddyfile = "service.testdomain.com {\n" +
-		"  proxy / https://service:5000/api\n" +
-		"}\n"
-
-	testGeneration(t, dockerClient, false, true, expectedCaddyfile, skipCaddyfileText)
-}
-
-func TestAddServiceWithMultipleConfigs(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ServicesData = []swarm.Service{
-		swarm.Service{
-			Spec: swarm.ServiceSpec{
-				Annotations: swarm.Annotations{
-					Name: "service",
-					Labels: map[string]string{
-						fmtLabel("%s_0.address"):            "service1.testdomain.com",
-						fmtLabel("%s_0.targetport"):         "5000",
-						fmtLabel("%s_0.targetpath"):         "/api",
-						fmtLabel("%s_0.proxy.health_check"): "/health",
-						fmtLabel("%s_0.proxy.transparent"):  "",
-						fmtLabel("%s_0.proxy.websocket"):    "",
-						fmtLabel("%s_0.basicauth"):          "/ user password",
-						fmtLabel("%s_0.tls.dns"):            "route53",
-						fmtLabel("%s_1.address"):            "service2.testdomain.com",
-						fmtLabel("%s_1.targetport"):         "5001",
-						fmtLabel("%s_1.tls.dns"):            "route53",
-					},
-				},
-			},
-			Endpoint: swarm.Endpoint{
-				VirtualIPs: []swarm.EndpointVirtualIP{
-					swarm.EndpointVirtualIP{
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-		},
-	}
-
-	const expectedCaddyfile = "service1.testdomain.com {\n" +
-		"  basicauth / user password\n" +
-		"  proxy / service:5000/api {\n" +
-		"    health_check /health\n" +
-		"    transparent\n" +
-		"    websocket\n" +
-		"  }\n" +
-		"  tls {\n" +
-		"    dns route53\n" +
-		"  }\n" +
-		"}\n" +
-		"service2.testdomain.com {\n" +
-		"  proxy / service:5001\n" +
-		"  tls {\n" +
-		"    dns route53\n" +
-		"  }\n" +
-		"}\n"
-
-	testGeneration(t, dockerClient, false, true, expectedCaddyfile, skipCaddyfileText)
-}
-
-func TestAddServiceProxyServiceTasks(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ServicesData = []swarm.Service{
-		swarm.Service{
-			Spec: swarm.ServiceSpec{
-				Annotations: swarm.Annotations{
-					Name: "service",
-					Labels: map[string]string{
-						fmtLabel("%s.address"):    "service.testdomain.com",
-						fmtLabel("%s.targetport"): "5000",
-					},
-				},
-			},
-			Endpoint: swarm.Endpoint{
-				VirtualIPs: []swarm.EndpointVirtualIP{
-					swarm.EndpointVirtualIP{
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-		},
-	}
-
-	const expectedCaddyfile = "service.testdomain.com {\n" +
-		"  proxy / tasks.service:5000\n" +
-		"}\n"
-
-	testGeneration(t, dockerClient, true, true, expectedCaddyfile, skipCaddyfileText)
-}
-
-func TestAddServiceMultipleAddresses(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ServicesData = []swarm.Service{
-		swarm.Service{
-			Spec: swarm.ServiceSpec{
-				Annotations: swarm.Annotations{
-					Name: "service",
-					Labels: map[string]string{
-						fmtLabel("%s.address"): "a.testdomain.com b.testdomain.com",
-					},
-				},
-			},
-			Endpoint: swarm.Endpoint{
-				VirtualIPs: []swarm.EndpointVirtualIP{
-					swarm.EndpointVirtualIP{
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-		},
-	}
-
-	const expectedCaddyfile = "a.testdomain.com b.testdomain.com {\n" +
-		"  proxy / service\n" +
-		"}\n"
-
-	testGeneration(t, dockerClient, false, true, expectedCaddyfile, skipCaddyfileText)
-}
-
-func TestAutomaticProxyDoesntOverrideCustomWithSameKey(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ServicesData = []swarm.Service{
-		swarm.Service{
-			Spec: swarm.ServiceSpec{
-				Annotations: swarm.Annotations{
-					Name: "service",
-					Labels: map[string]string{
-						fmtLabel("%s.address"): "testdomain.com",
-						fmtLabel("%s.proxy"):   "/ something",
-						fmtLabel("%s.proxy_1"): "/api external-api",
-					},
-				},
-			},
-			Endpoint: swarm.Endpoint{
-				VirtualIPs: []swarm.EndpointVirtualIP{
-					swarm.EndpointVirtualIP{
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-		},
-	}
-
-	const expectedCaddyfile = "testdomain.com {\n" +
-		"  proxy / something\n" +
-		"  proxy /api external-api\n" +
-		"}\n"
-
-	testGeneration(t, dockerClient, false, true, expectedCaddyfile, skipCaddyfileText)
-}
-
-func TestAddServiceFromDifferentNetwork(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ServicesData = []swarm.Service{
-		swarm.Service{
-			ID: "SERVICE-ID",
-			Spec: swarm.ServiceSpec{
-				Annotations: swarm.Annotations{
-					Name: "service",
-					Labels: map[string]string{
-						fmtLabel("%s.address"): "service.testdomain.com",
-					},
-				},
-			},
-			Endpoint: swarm.Endpoint{
-				VirtualIPs: []swarm.EndpointVirtualIP{
-					swarm.EndpointVirtualIP{
-						NetworkID: "other-network-id",
-					},
-				},
-			},
-		},
-	}
-
-	const expectedCaddyfile = ""
-
-	const expectedLogs = skipCaddyfileText +
-		"[ERROR] Service SERVICE-ID and caddy are not in same network\n"
-
-	testGeneration(t, dockerClient, false, true, expectedCaddyfile, expectedLogs)
-}
-
-func TestAddServiceFromDifferentNetworkSkipValidation(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ServicesData = []swarm.Service{
-		swarm.Service{
-			ID: "SERVICE-ID",
-			Spec: swarm.ServiceSpec{
-				Annotations: swarm.Annotations{
-					Name: "service",
-					Labels: map[string]string{
-						fmtLabel("%s.address"): "service.testdomain.com",
-					},
-				},
-			},
-			Endpoint: swarm.Endpoint{
-				VirtualIPs: []swarm.EndpointVirtualIP{
-					swarm.EndpointVirtualIP{
-						NetworkID: "other-network-id",
-					},
-				},
-			},
-		},
-	}
-
-	const expectedCaddyfile = "service.testdomain.com {\n" +
-		"  proxy / service\n" +
-		"}\n"
-
-	const expectedLogs = skipCaddyfileText
-
-	testGeneration(t, dockerClient, false, false, expectedCaddyfile, expectedLogs)
-}
-
-func TestAddServiceSwarmDisable(t *testing.T) {
-	dockerClient := createBasicDockerClientMock()
-	dockerClient.ServicesData = []swarm.Service{
-		swarm.Service{
-			ID: "SERVICE-ID",
-			Spec: swarm.ServiceSpec{
-				Annotations: swarm.Annotations{
-					Name: "service",
-					Labels: map[string]string{
-						fmtLabel("%s.address"): "service.testdomain.com",
-					},
-				},
-			},
-			Endpoint: swarm.Endpoint{
-				VirtualIPs: []swarm.EndpointVirtualIP{
-					swarm.EndpointVirtualIP{
-						NetworkID: caddyNetworkID,
-					},
-				},
-			},
-		},
-	}
-	dockerClient.InfoData = types.Info{
-		Swarm: swarm.Info{
-			LocalNodeState: swarm.LocalNodeStateInactive,
-		},
-	}
-
-	const expectedCaddyfile = ""
-
-	const expectedLogs = skipCaddyfileText +
-		"[INFO] Skipping services because swarm is not available\n" +
-		"[INFO] Skipping configs because swarm is not available\n"
-
-	testGeneration(t, dockerClient, false, true, expectedCaddyfile, expectedLogs)
 }
 
 func TestAddDockerConfigContent(t *testing.T) {
@@ -842,6 +146,7 @@ func createBasicDockerClientMock() *dockerClientMock {
 		ContainersData: []types.Container{},
 		ServicesData:   []swarm.Service{},
 		ConfigsData:    []swarm.Config{},
+		TasksData:      []swarm.Task{},
 		InfoData: types.Info{
 			Swarm: swarm.Info{
 				LocalNodeState: swarm.LocalNodeStateActive,
@@ -878,6 +183,7 @@ type dockerClientMock struct {
 	ContainersData       []types.Container
 	ServicesData         []swarm.Service
 	ConfigsData          []swarm.Config
+	TasksData            []swarm.Task
 	InfoData             types.Info
 	ContainerInspectData map[string]types.ContainerJSON
 	NetworkInspectData   map[string]types.NetworkResource
@@ -889,6 +195,20 @@ func (mock *dockerClientMock) ContainerList(ctx context.Context, options types.C
 
 func (mock *dockerClientMock) ServiceList(ctx context.Context, options types.ServiceListOptions) ([]swarm.Service, error) {
 	return mock.ServicesData, nil
+}
+
+func (mock *dockerClientMock) TaskList(ctx context.Context, options types.TaskListOptions) ([]swarm.Task, error) {
+	matchingTasks := []swarm.Task{}
+	for _, task := range mock.TasksData {
+		if !options.Filters.Match("service", task.ServiceID) {
+			continue
+		}
+		if !options.Filters.Match("desired-state", string(task.DesiredState)) {
+			continue
+		}
+		matchingTasks = append(matchingTasks, task)
+	}
+	return matchingTasks, nil
 }
 
 func (mock *dockerClientMock) Info(ctx context.Context) (types.Info, error) {
